fix: avoid torch.arange with int8 dtype (unsupported on NPU)
NPU aclnn does not support torch.arange with int8/uint8 dtype directly.
Use arange with default dtype then cast via .to(dtype).

+++ tests/test_allgather_batch.py
@@ -156,11 +156,11 @@ class TestNpuFunctional:
     @pytest.mark.parametrize("dtype", [torch.int8, torch.float16, torch.float32, torch.bfloat16])
     def test_single_desc(self, dtype):
         n = 128
-        inp = torch.arange(n, device=self.device, dtype=dtype) + self.rank * n
+        inp = (torch.arange(n, device=self.device) + self.rank * n).to(dtype)
         outs = torch.ops.custom_comm.allgather_batch([inp], self.hcom, self.world_size)
         assert outs[0].shape == (n * self.world_size,)
         for r in range(self.world_size):
-            expected = torch.arange(n, device=self.device, dtype=dtype) + r * n
+            expected = (torch.arange(n, device=self.device) + r * n).to(dtype)
             assert torch.equal(outs[0][r * n:(r + 1) * n], expected)
 
     def test_heterogeneous_int8_fp32(self):
@@ -213,7 +213,7 @@ class TestCcuPath:
     def test_ccu_matches_decomposed(self):
         """CCU output must match decomposed output bit-exactly."""
         import os
-        data = torch.arange(256, device=self.device, dtype=torch.int8) + self.rank
+        data = (torch.arange(256, device=self.device) + self.rank).to(torch.int8)
         scale = torch.randn(4, device=self.device, dtype=torch.float32)
 
         # Phase 1
